Added quarterly meter reset
This is still testing (my meter rolls over mid Feb 2023) and requires you to manually create four 'Date and/or time' helpers.
These need to be named "Gas - Quarter 1" through to number 4, with them being configured as input type 'date and time', and then edit it to set the date of your associated 'first day of the next billing quarter' (repeating for all four quarters). This creates the input_datetime.gas_quarter_1 (through to _4) that are called by the automation.